% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplicacao.R
\name{multiplicar}
\alias{multiplicar}
\title{Função Multiplicadora}
\usage{
multiplicar(a, b)
}
\arguments{
\item{a}{Um número.}

\item{b}{Outro número.}
}
\description{
Esta função serve para multiplicar dois números.
}
\examples{
multiplicar(1, 2)

}
